requirements.txt
streamlit
python-dotenv
azure-ai-documentintelligence
azure-storage-blob
